Move pageAttributeAssign and pageAttributeUnassign to page module
