+++ saleor/graphql/product/schema.py
@@ -49,8 +49,6 @@ from .mutations.attributes import (
     AttributeValueCreate,
     AttributeValueDelete,
     AttributeValueUpdate,
-    PageAttributeAssign,
-    PageAttributeUnassign,
     ProductAttributeAssign,
     ProductAttributeUnassign,
     ProductTypeReorderAttributes,
@@ -324,8 +322,6 @@ class ProductMutations(graphene.ObjectType):
     attribute_bulk_delete = AttributeBulkDelete.Field()
     product_attribute_assign = ProductAttributeAssign.Field()
     product_attribute_unassign = ProductAttributeUnassign.Field()
-    page_attribute_assign = PageAttributeAssign.Field()
-    page_attribute_unassign = PageAttributeUnassign.Field()
     attribute_update = AttributeUpdate.Field()
     attribute_translate = AttributeTranslate.Field()
     attribute_update_metadata = AttributeUpdateMeta.Field(
