+++ saleor/page/error_codes.py
@@ -8,3 +8,4 @@ class PageErrorCode(Enum):
     REQUIRED = "required"
     UNIQUE = "unique"
     DUPLICATED_INPUT_ITEM = "duplicated_input_item"
+    ATTRIBUTE_ALREADY_ASSIGNED = "attribute_already_assigned"
